Add default implementation to LibEvent::onThreadExit to avoid segfaults on exit
Sometimes on the server shutdown the pure virtual method LibEvent::onThreadExit
gets called. I'm not sure why this is happening, it seems like some race
condition is causing this method to be called after all the threads are stopped
and that is triggering the call to the pure virtual method.

As this is only a problem on the shutdown, we could live with this fix by now.