BE-862 Make get_called_class() consistent between HPHPI, HPHPand PHP.
In HPHPI, if a static method that calls get_called_class is called,
then get_called_class returns the name of the class using the case of
the invoker, instead of the name of the class. This is inconsistent
with behaviour in HPHP and PHP and causes singleton implementations to
generate multiple instances per class whenever class name uses
different case at invocation time and get_called_class() is used for
caching class instances.

E.g.:

class A { public static function a() { return get_called_class(); } }
echo A::a(); // 'A'
echo a::a(); // 'a'

which caused:

TuSubscriberApi::getInstance() !== TuSubscriberAPI::getInstance()

when using late static binding